fix(nfse): FK nfse_emissao_id INT casa com nfse_emissoes.increments (erro 3780) [M] (#2218)
A migration de cancelamento declarava nfse_emissao_id como unsignedBigInteger
(BIGINT), mas nfse_emissoes.id é criado via increments() (INT UNSIGNED) pela
migration NFSe 2026_05_01 — que roda antes e vence (a migration NfeBrasil
2026_05_11 tem guard hasTable e pula). FK com tipos diferentes → MySQL erro
3780 "incompatible", quebrando migrate:fresh e a suíte Pest inteira (Financeiro,
NFSe, etc).

Casa o tipo: unsignedBigInteger → unsignedInteger.

Bug pré-existente em main (não introduzido por nenhum PR de feature). Destrava
todo PR que roda a suíte com MySQL (ex: #2211) e o gate visual-regression.

Follow-up registrado: duas migrations criam nfse_emissoes com id de tipos
diferentes (NFSe increments/INT vs NfeBrasil id()/BIGINT) — drift entre módulos
a resolver separadamente.

+++ Modules/NfeBrasil/Database/Migrations/2026_05_12_120000_create_nfse_eventos_cancelamento_table.php
@@ -49,7 +49,10 @@ return new class extends Migration
         Schema::create('nfse_eventos_cancelamento', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('business_id')->index();
-            $table->unsignedBigInteger('nfse_emissao_id');
+            // INT (não BIGINT): casa com nfse_emissoes.id que é criado via
+            // increments() (INT UNSIGNED) pela migration NFSe 2026_05_01.
+            // unsignedBigInteger aqui gera FK incompatível (MySQL erro 3780).
+            $table->unsignedInteger('nfse_emissao_id');
             $table->string('driver_key', 32)
                 ->comment('ABRASF_V1, ABRASF_V2.04, GINFES, IPM, TIPLAN, NFSE_GOV_BR');
             $table->text('motivo')
